:hammer: Move JVM-bound implementation from commonMain to DesktopMain
