+++ app/src/commonMain/kotlin/com/crosspaste/ui/base/PasteTypeIconView.kt
@@ -3,7 +3,6 @@ package com.crosspaste.ui.base
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.contentColorFor
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.Dp
 import com.crosspaste.db.paste.PasteData
@@ -49,30 +48,6 @@ fun PasteTypeIconView(
             pasteData = pasteData,
             iconColor = iconColor,
             size = size,
-        )
-    }
-}
-
-@Composable
-fun SidePasteTypeIconView(
-    modifier: Modifier = Modifier,
-    pasteData: PasteData,
-    tint: Color,
-) {
-    pasteData.source?.let {
-        SideAppSourceIcon(
-            modifier = modifier,
-            pasteData = pasteData,
-        ) {
-            SideDefaultPasteTypeIcon(
-                pasteData = pasteData,
-                tint = tint,
-            )
-        }
-    } ?: run {
-        SideDefaultPasteTypeIcon(
-            pasteData = pasteData,
-            tint = tint,
         )
     }
 }
